The original XFA version was correct, and the master version here is
wrong. The two versions are now in sync, but incorrect.
So we need to fix this here and then merge to XFA again.

Also fix what are now incorrect uses of CountAnnots() and do some
cleanups.

BUG=543049
TBR=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1413973011 .

(cherry picked from commit c91d5c276b6b986f645d34a0e583fbe4f73e74bc)

Review URL: https://codereview.chromium.org/1422503012 .

+++ fpdfsdk/include/fsdk_mgr.h
@@ -557,7 +557,7 @@ class CPDFSDK_PageView final {
   CPDFSDK_Annot* AddAnnot(CPDF_Annot* pPDFAnnot);
   CPDFSDK_Annot* AddAnnot(IXFA_Widget* pPDFAnnot);
   FX_BOOL DeleteAnnot(CPDFSDK_Annot* pAnnot);
-  int CountAnnots() const;
+  size_t CountAnnots() const;
   CPDFSDK_Annot* GetAnnot(size_t nIndex);
   CPDFSDK_Annot* GetAnnotByDict(CPDF_Dictionary* pDict);
   CPDFSDK_Annot* GetAnnotByXFAWidget(IXFA_Widget* hWidget);
